pyproject.toml
[tool.poetry]
name = "yolo_demo"
version = "0.12.0"
description = ""
authors = ["Angry Fennec <shellfix@gmail.com>"]

[tool.poetry.dependencies]
python = "~=3.9"
numpy = "^1.22.3"
opencv-python = "^4.5.5"
Pillow = "^9.0.1"
tensorflow = "2.8.0"
easydict = "^1.9"
streamlit = "^1.8.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pre-commit = "^2.18.1"
flake8 = "^4.0.1"
black = "^22.3.0"

[flake8]
ignore = "D203"
exclude = ".git, dist, ./yolo_demo/core"
max-complexity = 10

[tool.black]
line-length = 79
exclude = ".git, dist, yolo_demo/core"
target-version = ['py39']

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
